create AssingmentExpr class
