+++ src/astx/tools/transpilers/python.py
@@ -413,11 +413,6 @@ class ASTxPythonTranspiler:
         return node.name
 
     @dispatch  # type: ignore[no-redef]
-    def visit(self, node: astx.WalrusOp) -> str:
-        """Handle Walrus operator."""
-        return f"({self.visit(node.lhs)} := {self.visit(node.rhs)})"
-
-    @dispatch  # type: ignore[no-redef]
     def visit(self, node: astx.VariableAssignment) -> str:
         """Handle VariableAssignment nodes."""
         target = node.name
@@ -429,6 +424,11 @@ class ASTxPythonTranspiler:
         """Handle VariableDeclaration nodes."""
         value = self.visit(node.value)
         return f"{node.name}: {node.value.type_.__class__.__name__} = {value}"
+
+    @dispatch  # type: ignore[no-redef]
+    def visit(self, node: astx.WalrusOp) -> str:
+        """Handle Walrus operator."""
+        return f"({self.visit(node.lhs)} := {self.visit(node.rhs)})"
 
     @dispatch  # type: ignore[no-redef]
     def visit(self, node: astx.WhileExpr) -> str:
